requirements.txt
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-huggingface
sentence-transformers
faiss-cpu
mcp
python-dotenv
google-generativeai